move generated code to macro
