+++ examples/todomvc/global-state-ecs/global_state.rs
@@ -2,10 +2,8 @@ use crate::component::*;
 use bevy::prelude::*;
 use bevy_dioxus::core::prelude::*;
 use chrono::{DateTime, Utc};
-use dioxus::fermi::{Atom, AtomRoot, Readable};
 
-// #[derive(GlobalState)]
-#[derive(Clone, Debug)]
+#[derive(GlobalStateCommand, Clone, Debug)]
 pub enum GlobalStateCommand {
     TodoList(Vec<UiTodo>),
 }
@@ -37,50 +35,50 @@ impl From<(Entity, &Title, Option<&DoneAt>, &Timestamp)> for UiTodo {
 }
 
 // Generated
-use bevy_dioxus::desktop::event::VDomCommand;
-use futures_intrusive::channel::{shared::Sender, TrySendError};
-use std::rc::Rc;
+// use bevy_dioxus::desktop::event::VDomCommand;
+// use futures_intrusive::channel::{shared::Sender, TrySendError};
+// use std::rc::Rc;
 
-pub static TODO_LIST: Atom<Vec<UiTodo>> = |_| vec![];
+// pub static TODO_LIST: Atom<Vec<UiTodo>> = |_| vec![];
 
-impl GlobalStateHandler for GlobalStateCommand {
-    fn handler(self, root: Rc<AtomRoot>) {
-        match self {
-            GlobalStateCommand::TodoList(x) => root.set(TODO_LIST.unique_id(), x),
-        }
-    }
-}
+// impl GlobalStateHandler for GlobalStateCommand {
+//     fn handler(self, root: Rc<AtomRoot>) {
+//         match self {
+//             GlobalStateCommand::TodoList(x) => root.set(TODO_LIST.unique_id(), x),
+//         }
+//     }
+// }
 
-pub struct GlobalStatePlugin;
+// pub struct GlobalStatePlugin;
 
-impl Plugin for GlobalStatePlugin {
-    fn build(&self, app: &mut App) {
-        app.add_event::<GlobalStateCommand>()
-            .add_system(apply_global_state_command);
-    }
-}
+// impl Plugin for GlobalStatePlugin {
+//     fn build(&self, app: &mut App) {
+//         app.add_event::<GlobalStateCommand>()
+//             .add_system(apply_global_state_command);
+//     }
+// }
 
-fn apply_global_state_command(
-    mut events: EventReader<GlobalStateCommand>,
-    vdom_tx: Res<Sender<VDomCommand<GlobalStateCommand>>>,
-) {
-    for e in events.iter() {
-        match vdom_tx.try_send(VDomCommand::GlobalState(e.clone())) {
-            Ok(()) => {}
-            Err(e) => match e {
-                TrySendError::Full(e) => {
-                    error!(
-                        "Failed to send VDomCommand: channel is full: event: {:?}",
-                        e
-                    );
-                }
-                TrySendError::Closed(e) => {
-                    error!(
-                        "Failed to send VDomCommand: channel is closed: event: {:?}",
-                        e
-                    );
-                }
-            },
-        }
-    }
-}
+// fn apply_global_state_command(
+//     mut events: EventReader<GlobalStateCommand>,
+//     vdom_tx: Res<Sender<VDomCommand<GlobalStateCommand>>>,
+// ) {
+//     for e in events.iter() {
+//         match vdom_tx.try_send(VDomCommand::GlobalState(e.clone())) {
+//             Ok(()) => {}
+//             Err(e) => match e {
+//                 TrySendError::Full(e) => {
+//                     error!(
+//                         "Failed to send VDomCommand: channel is full: event: {:?}",
+//                         e
+//                     );
+//                 }
+//                 TrySendError::Closed(e) => {
+//                     error!(
+//                         "Failed to send VDomCommand: channel is closed: event: {:?}",
+//                         e
+//                     );
+//                 }
+//             },
+//         }
+//     }
+// }
